Removes React anti pattern + nicer error handling.
fetch-gif.js used to get a reference to react component and call `setState` on it.
It's an anti pattern in react world. It's better to have `this.setState` within
a component itself. It was not very clear for students where imageURL comes from
and was a bit difficult for volunteers to explain it due students'
lack of knowledge of JS function context etc.